refactor(BarChart,BarStack): rewrite stories
