+++ src/components/BarChart/__stories__/Template.stories.tsx
@@ -1,0 +1,7 @@
+import { ComponentStory } from '@storybook/react'
+import BarChart from '..'
+import { barChartSimpleData } from './mockData'
+
+export const Template: ComponentStory<typeof BarChart> = ({ ...props }) => (
+  <BarChart data={barChartSimpleData} {...props} />
+)
