+++ src/components/BarChart/__stories__/index.stories.tsx
@@ -1,89 +1,19 @@
-import { Meta, Story } from '@storybook/react'
-import { format } from 'date-fns'
-import { ComponentProps } from 'react'
+import { ComponentMeta } from '@storybook/react'
 import BarChart from '..'
-import {
-  barChartMultiData,
-  barChartPositiveNegativeData,
-  barChartSimpleData,
-} from './mockData'
 
 export default {
   component: BarChart,
+  parameters: {
+    docs: {
+      description: {
+        component: 'BarChart is a chart component that renders bars.',
+      },
+    },
+  },
   title: 'Components/Data Display/Chart/BarChart',
-} as Meta
+} as ComponentMeta<typeof BarChart>
 
-const Template: Story<ComponentProps<typeof BarChart>> = args => (
-  <BarChart data={barChartSimpleData} {...args} />
-)
-
-export const Default = Template.bind({})
-
-export const FormattedValuesAndTooltip = Template.bind({})
-FormattedValuesAndTooltip.decorators = [
-  () => (
-    <BarChart
-      data={barChartSimpleData}
-      axisFormatters={{
-        bottom: value => format(new Date(value), 'dd-MM-Y'),
-        left: value => `${value.toString()} kb`,
-      }}
-      tooltipFunction={({ value, indexValue, color }) => ({
-        color,
-        formattedValue: `${value} kb`,
-        indexValue: format(new Date(indexValue), 'dd-MM-Y'),
-      })}
-    />
-  ),
-]
-
-export const MultiSeries = Template.bind({})
-MultiSeries.decorators = [
-  () => (
-    <BarChart
-      keys={['sent', 'received']}
-      data={barChartMultiData}
-      axisFormatters={{
-        bottom: value => format(new Date(value), 'dd-MM-Y'),
-      }}
-      tooltipFunction={({ value, indexValue, color }) => ({
-        color,
-        formattedValue: `${value}`,
-        indexValue: format(new Date(indexValue), 'dd-MM-Y'),
-      })}
-    />
-  ),
-]
-
-export const PositiveNegative = Template.bind({})
-PositiveNegative.decorators = [
-  () => (
-    <BarChart
-      height={200}
-      data={barChartPositiveNegativeData}
-      axisFormatters={{
-        bottom: value => format(new Date(value), 'dd-MM-Y'),
-        left: value => {
-          if (value === 1) return 'Active'
-          if (value === -1) return 'Inactive'
-
-          return ''
-        },
-      }}
-      tooltipFunction={({ value, indexValue, color }) => ({
-        color,
-        formattedValue: value === 1 ? 'Active' : 'Inactive',
-        indexValue: format(new Date(indexValue), 'dd-MM-Y'),
-      })}
-      tickValues={{
-        left: 10,
-      }}
-      chartProps={{
-        colors: ({ value }) => (value === 1 ? 'green' : 'red'),
-        gridYValues: [-1, 0, 1],
-        maxValue: 1,
-        minValue: -1,
-      }}
-    />
-  ),
-]
+export { Playground } from './Playground.stories'
+export { FormattedValuesAndTooltip } from './FormattedValuesAndTooltip.stories'
+export { MultiSeries } from './MultiSeries.stories'
+export { PositiveNegative } from './PositiveNegative.stories'
