+++ src/components/BarStack/__stories__/Template.stories.tsx
@@ -1,0 +1,7 @@
+import { ComponentStory } from '@storybook/react'
+import BarStack from '..'
+import { fakeData } from './mockData'
+
+export const Template: ComponentStory<typeof BarStack> = ({ ...props }) => (
+  <BarStack {...props} data={fakeData} />
+)
