+++ src/components/BarStack/__stories__/WithMax.stories.tsx
@@ -1,0 +1,7 @@
+import { Template } from './Template.stories'
+
+export const WithMax = Template.bind({})
+
+WithMax.args = {
+  total: 500,
+}
