+++ src/components/BarStack/__stories__/mockData.ts
@@ -1,0 +1,13 @@
+export const fakeData = [
+  { id: '1', text: 'Hello', tooltip: 'Hello', value: 20 },
+  { id: '2', text: 'Nice component', tooltip: 'Nice component', value: 42 },
+  { id: '3', text: 'Encore', tooltip: 'Encore', value: 42 },
+  { id: '4', text: 'Next', tooltip: 'Next', value: 42 },
+  {
+    id: '5',
+    text: 'Next should be same as first',
+    tooltip: 'Next should be same as first',
+    value: 42,
+  },
+  { id: '6', text: 'Bye bye', tooltip: 'Bye bye', value: 42 },
+]
